Move bottom nav screens to `(tabs)` for use with `expo-router` <Tabs>
